Port to AWS Go SDK v2
v1 is deprecated, and support will end a little over a year from now.
This ports the route53 plugin to v2.

The major difference is that listing resource records no longer uses a
callback function; the results are retrieved directly, but paging has
to be handled by the caller.

This hasn’t been tested on AWS yet.

Signed-off-by: Stephen Kitt <skitt@redhat.com>